move createZimWikiHeaderAndTitleContents method
